requirements.txt
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
telebot-calendar==1.2
peewee==3.15.2
loguru==0.7.0
requests==2.28.2